Split world state surface
